pyproject.toml
[tool.poetry]
name = "websocket-ping-pong"
version = "0.1.0"
description = "Secure ping-pong server and client using websockets"
authors = ["Nikolas <mmmsvit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
websockets = "==10.4"
zeroconf = "==0.47.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
tornado = "^6.2.0"
websocket-client = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
